Tolerate gardener critical components unavailable
Ideally, we make that configurable, but I do not see how.
Having a toleration, which doesn't exist doesn't hurt.

This taint will only be removed after we have managed to start
a VM, so we need this agent to run with this taint.

We still do not tolerate the following taints:
  node.kubernetes.io/not-ready
  node.kubernetes.io/network-unavailable

So, the node is ready and working, except for our test.